bug fix: box sizing for all elements
Justifying text revealed considerable overflow on the right side of text
columns. I suspect that this is simply an issue with box sizing, but we will
find out through EXPERIMENTATION!

So far it just seems to be a funky way that the chat window paints the HTML.
Some padding is the quick fix to keep all of the letters legible.